# Optimised Bytes Array: Remove non-vectorised API + Adopt Numpy Arrays