--- a/vernamveil/_hash_utils.py
+++ b/vernamveil/_hash_utils.py
@@ -207,7 +207,7 @@ def fold_bytes_to_uint64(
 
 def hash_numpy(
     i: np.ndarray[tuple[int], np.dtype[np.uint64]],
-    seed: bytes | None = None,
+    seed: bytes | bytearray | None = None,
     hash_name: HashType = "blake2b",
     hash_size: int | None = None,
 ) -> np.ndarray[tuple[int, int], np.dtype[np.uint8]]:
@@ -221,7 +221,7 @@ def hash_numpy(
 
     Args:
         i (np.ndarray[tuple[int], np.dtype[np.uint64]]): NumPy array of indices (dtype should be unsigned 64-bit integer).
-        seed (bytes, optional): The seed bytes are prepended to the index. If None, hashes only the index.
+        seed (bytes or bytearray, optional): The seed bytes are prepended to the index. If None, hashes only the index.
         hash_name (HashType): Hash function to use ("blake2b", "blake3" or "sha256"). The blake3 is only available
             if the C extension is installed. Defaults to "blake2b".
         hash_size (int, optional): Size of the hash output in bytes. Should be 64 for blake2b, larger than 0 for blake3