--- a/vernamveil/_types.py
+++ b/vernamveil/_types.py
@@ -2,25 +2,8 @@
 
 from typing import Any, Literal
 
-# Vectorisation types and imports
-np: Any
-_Integer: Any
-_Bytes: Any
-try:
-    import numpy
-
-    np = numpy
-    _Integer = int | np.ndarray[np.uint64]
-    _Bytes = bytes | np.ndarray[np.uint8]
-    _HAS_NUMPY = True
-except ImportError:
-    np = None
-    _Integer = int
-    _Bytes = bytes
-    _HAS_NUMPY = False
-
 # C module types and imports
-_HashType: Any
+HashType: Any
 _bytesearchffi: Any
 _npblake2bffi: Any
 _npblake3ffi: Any
@@ -29,7 +12,7 @@ try:
     from nphash import _bytesearchffi, _npblake2bffi, _npblake3ffi, _npsha256ffi
 
     _HAS_C_MODULE = True
-    _HashType = Literal["blake2b", "blake3", "sha256"]
+    HashType = Literal["blake2b", "blake3", "sha256"]
 except ImportError:
     _bytesearchffi = None
     _npblake2bffi = None
@@ -37,4 +20,4 @@ except ImportError:
     _npsha256ffi = None
 
     _HAS_C_MODULE = False
-    _HashType = Literal["blake2b", "sha256"]
+    HashType = Literal["blake2b", "sha256"]