--- a/vernamveil/_vernamveil.py
+++ b/vernamveil/_vernamveil.py
@@ -296,7 +296,7 @@ class VernamVeil(_Cypher):
         exact_size = pad_count * self._delimiter_size + random_size + message_len
 
         # Build the noisy message by combining fake and shuffled real chunks
-        noisy_blocks = bytearray(exact_size)
+        noisy_blocks = np.empty(exact_size, dtype=np.uint8)
         current_rec_loc = 0
         current_rnd_loc = 0
         for chunk_range in shuffled_chunk_ranges:
@@ -333,18 +333,18 @@ class VernamVeil(_Cypher):
                 noisy_blocks[current_rec_loc:next_rec_loc] = part
                 current_rec_loc = next_rec_loc
 
-        return memoryview(noisy_blocks)
+        return noisy_blocks.data
 
-    def _deobfuscate(self, noisy: bytearray, seed: bytes, delimiter: memoryview) -> bytearray:
+    def _deobfuscate(self, noisy: memoryview, seed: bytes, delimiter: memoryview) -> memoryview:
         """Remove noise and extract real chunks from a shuffled noisy message.
 
         Args:
-            noisy (bytearray): Decrypted and obfuscated message.
+            noisy (memoryview): Decrypted and obfuscated message.
             seed (bytes): Seed for deterministic chunk deshuffling.
             delimiter (memoryview): Delimiter used to detect chunks.
 
         Returns:
-            bytearray: Original message reconstructed from real chunks.
+            memoryview: Original message reconstructed from real chunks.
         """
         # Estimate the ranges of all chunks
         delimiter_len = len(delimiter)
@@ -381,21 +381,20 @@ class VernamVeil(_Cypher):
         )
 
         # Reconstruct and unshuffle the message
-        message = bytearray(exact_size)
-        view = memoryview(noisy)
+        message = np.empty(exact_size, dtype=np.uint8)
         current_loc = 0
         for pos in shuffled_positions:
             start, end = all_chunk_ranges[pos]
 
             next_loc = current_loc + end - start
-            message[current_loc:next_loc] = view[start:end]
+            message[current_loc:next_loc] = noisy[start:end]
             current_loc = next_loc
 
-        return message
+        return message.data
 
     def _xor_with_key(
         self, data: memoryview, seed: bytes, is_encode: bool
-    ) -> tuple[bytearray, bytes]:
+    ) -> tuple[memoryview, bytes]:
         """Encrypt or decrypt data using XOR with the generated keystream.
 
         Args:
@@ -404,14 +403,11 @@ class VernamVeil(_Cypher):
             is_encode (bool): True for encryption, False for decryption.
 
         Returns:
-            tuple[bytearray, bytes]: Processed data and the final seed.
+            tuple[memoryview, bytes]: Processed data and the final seed.
         """
-        # Preallocate memory and avoid copying when slicing
+        # Preallocate memory
         data_len = len(data)
-        result = bytearray(data_len)
-
-        # Create a numpy array on top of the bytearray to vectorise and still have access to original bytearray
-        processed = np.frombuffer(result, dtype=np.uint8)
+        result = np.empty(data_len, dtype=np.uint8)
 
         for start, end in self._generate_chunk_ranges(data_len):
             # Generate a key using fx
@@ -421,7 +417,8 @@ class VernamVeil(_Cypher):
             # XOR the chunk with the key
             # Store the slicing to avoid duplicate ops
             data_slice = data[start:end]
-            processed_slice = processed[start:end]
+            processed_slice = result[start:end]
+
             # Writing to slices modifies the original data
             np.bitwise_xor(data_slice, keystream, out=processed_slice)
             plaintext_data = data_slice if is_encode else processed_slice.data
@@ -429,7 +426,7 @@ class VernamVeil(_Cypher):
             # Refresh the seed differently for encoding and decoding
             seed = self._hash(seed, [plaintext_data])
 
-        return result, seed
+        return result.data, seed
 
     def _generate_delimiter(self, seed: bytes) -> tuple[memoryview, bytes]:
         """Create a delimiter sequence using the key stream and update the seed.
@@ -446,7 +443,7 @@ class VernamVeil(_Cypher):
 
     def encode(
         self, message: bytes | bytearray | memoryview, seed: bytes
-    ) -> tuple[bytearray, bytes]:
+    ) -> tuple[memoryview, bytes]:
         """Encrypt a message.
 
         Args:
@@ -454,7 +451,7 @@ class VernamVeil(_Cypher):
             seed (bytes): Initial seed for encryption.
 
         Returns:
-            tuple[bytearray, bytes]: Encrypted message and final seed.
+            tuple[memoryview, bytes]: Encrypted message and final seed.
 
         Raises:
             ValueError: If the delimiter appears in the message.
@@ -464,7 +461,7 @@ class VernamVeil(_Cypher):
             message = memoryview(message)
 
         # Store the output parts
-        output: list[bytes | bytearray] = []
+        output: list[memoryview] = []
 
         # SIV seed initialisation: Encrypt and prepend a synthetic IV (SIV) derived from the seed and message.
         # This prevents deterministic keystreams on the first block and makes the scheme resilient to seed reuse.
@@ -507,21 +504,21 @@ class VernamVeil(_Cypher):
         if self._auth_encrypt:
             # The tag is computed over the configuration of the cypher and the cyphertext.
             tag = self._hash(auth_seed, [str(self).encode(), cyphertext], use_hmac=True)
-            output.append(tag)
+            output.append(memoryview(tag))
 
         # Concatenate all parts into a single bytearray
-        result = bytearray(sum(len(part) for part in output))
+        result = np.empty(sum(len(part) for part in output), dtype=np.uint8)
         current_loc = 0
         for part in output:
             next_loc = current_loc + len(part)
             result[current_loc:next_loc] = part
             current_loc = next_loc
 
-        return result, last_seed
+        return result.data, last_seed
 
     def decode(
         self, cyphertext: bytes | bytearray | memoryview, seed: bytes
-    ) -> tuple[bytearray, bytes]:
+    ) -> tuple[memoryview, bytes]:
         """Decrypt an encoded message.
 
         Args:
@@ -529,7 +526,7 @@ class VernamVeil(_Cypher):
             seed (bytes): Initial seed for decryption.
 
         Returns:
-            tuple[bytearray, bytes]: Decrypted message and final seed.
+            tuple[memoryview, bytes]: Decrypted message and final seed.
 
         Raises:
             ValueError: If the authentication tag does not match.