Fixes https://github.com/datumbox/VernamVeil/issues/41

This PR eliminates the scalar API and fully embraces numpy as a core dependency. The introduces np.arrays as the key buffer of the cypher (to replace bytearrays). Due to the vectorised nature of Numpy, this makes the slicing, accessing and allocation much much faster.

Previous baseline from #40:
```
The 'encode' step took 4.313 seconds.
The 'decode' step took 4.183 seconds.
```

Using (782fa1d) we have 20-23% faster:
```
The 'encode' step took 3.307 seconds.
The 'decode' step took 3.356 seconds.
```