@datumbox 2025-06-09
Closing for now as we went with #43